+++ Museum.App.Services/Implementation/Repositories/GalleryObjectRepository.cs
@@ -1,11 +1,10 @@
-using Museum.App.Services.Attributes;
 using Museum.App.Services.Interfaces.Repositories;
 using Museum.App.ViewModels.GalleryObject;
-using Dapper;
+using Museum.Models.GalleryObjectModels;
+using Museum.App.Services.Attributes;
 using Microsoft.Data.SqlClient;
 using System.Data;
-using Museum.App.Services.Abstractions;
-using Museum.Models.TableModels;
+using Dapper;
 
 namespace Museum.App.Services.Implementation.Repositories
 {
@@ -19,22 +18,30 @@ namespace Museum.App.Services.Implementation.Repositories
             _connection = connection;
         }
 
-        public GalleryMainSectionImages GetGalleryObjectImages(int Object_ID)
+        public string GetGalleryDecsAsString(int Object_ID)
         {
-            using var sql = new SqlConnection(_connection);
-            return sql.QuerySingleOrDefault<GalleryMainSectionImages>(
-                "GetGalleryObjectImages", 
-                new { ObjectID = Object_ID }, 
-                commandType: CommandType.StoredProcedure);
+            return GetFromObject<string>(Object_ID, "GetObjectDetailsAsString");
         }
 
-        public GalleryUlViewModel GetGalleryUl(int Object_ID)
+        public ObjectDetailsModel GetGalleryDecsAsTable(int Object_ID)
+        {
+            return GetFromObject<ObjectDetailsModel>(Object_ID, "GetObjectDetailsAsModel");
+        }
+
+        public GalleryMainSectionImages GetGalleryObjectImages(int Object_ID)
+        {
+            return GetFromObject<GalleryMainSectionImages>(Object_ID, "GetGalleryObjectImages");
+        }
+
+        public GalleryUlModel GetGalleryUl(int Object_ID)
+        {
+            return GetFromObject<GalleryUlModel>(Object_ID, "GetGalleryObjectDescByID");
+        }
+
+        private T GetFromObject<T>(int Object_ID, string ProcedureName)
         {
             using var sql = new SqlConnection(_connection);
-            return sql.QuerySingleOrDefault<GalleryUlViewModel>(
-                "GetGalleryObjectDescByID",
-                new { ObjectID = Object_ID }, 
-                commandType: CommandType.StoredProcedure);
+            return sql.QuerySingleOrDefault<T>(ProcedureName, new { ObjectID = Object_ID }, commandType: CommandType.StoredProcedure);
         }
     }
 }
