--- Museum.App.Services/Implementation/Repositories/RoleRepository.cs
@@ -1,123 +1,0 @@
-﻿using Museum.Models.TableModels;
-using System;
-using System.Collections.Generic;
-using System.Data.SqlClient;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
-using Dapper;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.Data.SqlClient; 
-
-namespace Museum.App.Services.Implementation.Repositories
-{
-    public class RoleRepository : IRoleStore<ApplicationRole>
-    {
-        private readonly string _connectionString;
-
-        public RoleRepository(string connectionString)
-        {
-            _connectionString = connectionString;
-        }
-
-        public async Task<IdentityResult> CreateAsync(ApplicationRole role, CancellationToken cancellationToken)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                await connection.OpenAsync(cancellationToken);
-                role.ApplicationRole_ID = await connection.QuerySingleAsync<int>($@"INSERT INTO [ApplicationRole] ([Name], [NormalizedName])
-                    VALUES (@{nameof(ApplicationRole.Name)}, @{nameof(ApplicationRole.NormalizedName)});
-                    SELECT CAST(SCOPE_IDENTITY() as int)", role);
-            }
-
-            return IdentityResult.Success;
-        }
-
-        public async Task<IdentityResult> UpdateAsync(ApplicationRole role, CancellationToken cancellationToken)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                await connection.OpenAsync(cancellationToken);
-                await connection.ExecuteAsync($@"UPDATE [ApplicationRole] SET
-                    [Name] = @{nameof(ApplicationRole.Name)},
-                    [NormalizedName] = @{nameof(ApplicationRole.NormalizedName)}
-                    WHERE [ApplicationRole_ID] = @{nameof(ApplicationRole.ApplicationRole_ID)}", role);
-            }
-
-            return IdentityResult.Success;
-        }
-
-        public async Task<IdentityResult> DeleteAsync(ApplicationRole role, CancellationToken cancellationToken)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                await connection.OpenAsync(cancellationToken);
-                await connection.ExecuteAsync($"DELETE FROM [ApplicationRole] WHERE [ApplicationRole_ID] = @{nameof(ApplicationRole.ApplicationRole_ID)}", role);
-            }
-
-            return IdentityResult.Success;
-        }
-
-        public Task<string> GetRoleIdAsync(ApplicationRole role, CancellationToken cancellationToken)
-        {
-            return Task.FromResult(role.ApplicationRole_ID.ToString());
-        }
-
-        public Task<string> GetRoleNameAsync(ApplicationRole role, CancellationToken cancellationToken)
-        {
-            return Task.FromResult(role.Name);
-        }
-
-        public Task SetRoleNameAsync(ApplicationRole role, string roleName, CancellationToken cancellationToken)
-        {
-            role.Name = roleName;
-            return Task.FromResult(0);
-        }
-
-        public Task<string> GetNormalizedRoleNameAsync(ApplicationRole role, CancellationToken cancellationToken)
-        {
-            return Task.FromResult(role.NormalizedName);
-        }
-
-        public Task SetNormalizedRoleNameAsync(ApplicationRole role, string normalizedName, CancellationToken cancellationToken)
-        {
-            role.NormalizedName = normalizedName;
-            return Task.FromResult(0);
-        }
-
-        public async Task<ApplicationRole> FindByIdAsync(string roleId, CancellationToken cancellationToken)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                await connection.OpenAsync(cancellationToken);
-                return await connection.QuerySingleOrDefaultAsync<ApplicationRole>($@"SELECT * FROM [ApplicationRole]
-                    WHERE [ApplicationRole_ID] = @{nameof(roleId)}", new { roleId });
-            }
-        }
-
-        public async Task<ApplicationRole> FindByNameAsync(string normalizedRoleName, CancellationToken cancellationToken)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                await connection.OpenAsync(cancellationToken);
-                return await connection.QuerySingleOrDefaultAsync<ApplicationRole>($@"SELECT * FROM [ApplicationRole]
-                    WHERE [NormalizedName] = @{nameof(normalizedRoleName)}", new { normalizedRoleName });
-            }
-        }
-
-        public void Dispose()
-        {
-            // Nothing to dispose.
-        }
-    }
-}
