--- Museum.App.Services/Implementation/Repositories/UserRepository.cs
@@ -1,309 +1,0 @@
-﻿using Museum.App.Services.Interfaces.Repositories;
-using Museum.Models.TableModels;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.Data.SqlClient;
-using Dapper;
-using Museum.App.Services.Abstractions;
-
-namespace Museum.App.Services.Implementation.Repositories
-{
-    public class UserRepository : IUserRepository<ApplicationUser>
-    {
-        private readonly string _connectionString;
-
-        public UserRepository(string connectionString)
-        {
-            _connectionString = connectionString;
-        }
-
-        public async Task<IdentityResult> CreateAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                await connection.OpenAsync(cancellationToken);
-                user.ApplicationUser_ID = await connection.QuerySingleAsync<int>($@"INSERT INTO [ApplicationUser] ([UserName], [NormalizedUserName], [Email],
-                    [NormalizedEmail], [EmailConfirmed], [PasswordHash], [PhoneNumber], [PhoneNumberConfirmed], [TwoFactorEnabled])
-                    VALUES (@{nameof(ApplicationUser.UserName)}, @{nameof(ApplicationUser.NormalizedUserName)}, @{nameof(ApplicationUser.Email)},
-                    @{nameof(ApplicationUser.NormalizedEmail)}, @{nameof(ApplicationUser.EmailConfirmed)}, @{nameof(ApplicationUser.PasswordHash)},
-                    @{nameof(ApplicationUser.PhoneNumber)}, @{nameof(ApplicationUser.PhoneNumberConfirmed)}, @{nameof(ApplicationUser.TwoFactorEnabled)});
-                    SELECT CAST(SCOPE_IDENTITY() as int)", user);
-            }
-
-            return IdentityResult.Success;
-        }
-
-        public async Task<IdentityResult> DeleteAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                await connection.OpenAsync(cancellationToken);
-                await connection.ExecuteAsync($"DELETE FROM [ApplicationUser] WHERE [ApplicationUser_ID] = @{nameof(ApplicationUser.ApplicationUser_ID)}", user);
-            }
-
-            return IdentityResult.Success;
-        }
-
-        public async Task<ApplicationUser> FindByIdAsync(string userId, CancellationToken cancellationToken)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                await connection.OpenAsync(cancellationToken);
-                return await connection.QuerySingleOrDefaultAsync<ApplicationUser>($@"SELECT * FROM [ApplicationUser]
-                    WHERE [ApplicationUser_ID] = @{nameof(userId)}", new { userId });
-            }
-        }
-
-        public async Task<ApplicationUser> FindByNameAsync(string normalizedUserName, CancellationToken cancellationToken)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                await connection.OpenAsync(cancellationToken);
-                return await connection.QuerySingleOrDefaultAsync<ApplicationUser>($@"SELECT * FROM [ApplicationUser]
-                    WHERE [NormalizedUserName] = @{nameof(normalizedUserName)}", new { normalizedUserName });
-            }
-        }
-
-        public Task<string> GetNormalizedUserNameAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            return Task.FromResult(user.NormalizedUserName);
-        }
-
-        public Task<string> GetUserIdAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            return Task.FromResult(user.ApplicationUser_ID.ToString());
-        }
-
-        public Task<string> GetUserNameAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            return Task.FromResult(user.UserName);
-        }
-
-        public Task SetNormalizedUserNameAsync(ApplicationUser user, string normalizedName, CancellationToken cancellationToken)
-        {
-            user.NormalizedUserName = normalizedName;
-            return Task.FromResult(0);
-        }
-
-        public Task SetUserNameAsync(ApplicationUser user, string userName, CancellationToken cancellationToken)
-        {
-            user.UserName = userName;
-            return Task.FromResult(0);
-        }
-
-        public async Task<IdentityResult> UpdateAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                await connection.OpenAsync(cancellationToken);
-                await connection.ExecuteAsync($@"UPDATE [ApplicationUser] SET
-                    [UserName] = @{nameof(ApplicationUser.UserName)},
-                    [NormalizedUserName] = @{nameof(ApplicationUser.NormalizedUserName)},
-                    [Email] = @{nameof(ApplicationUser.Email)},
-                    [NormalizedEmail] = @{nameof(ApplicationUser.NormalizedEmail)},
-                    [EmailConfirmed] = @{nameof(ApplicationUser.EmailConfirmed)},
-                    [PasswordHash] = @{nameof(ApplicationUser.PasswordHash)},
-                    [PhoneNumber] = @{nameof(ApplicationUser.PhoneNumber)},
-                    [PhoneNumberConfirmed] = @{nameof(ApplicationUser.PhoneNumberConfirmed)},
-                    [TwoFactorEnabled] = @{nameof(ApplicationUser.TwoFactorEnabled)}
-                    WHERE [ApplicationUser_ID] = @{nameof(ApplicationUser.ApplicationUser_ID)}", user);
-            }
-
-            return IdentityResult.Success;
-        }
-
-        public Task SetEmailAsync(ApplicationUser user, string email, CancellationToken cancellationToken)
-        {
-            user.Email = email;
-            return Task.FromResult(0);
-        }
-
-        public Task<string> GetEmailAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            return Task.FromResult(user.Email);
-        }
-
-        public Task<bool> GetEmailConfirmedAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            return Task.FromResult(user.EmailConfirmed);
-        }
-
-        public Task SetEmailConfirmedAsync(ApplicationUser user, bool confirmed, CancellationToken cancellationToken)
-        {
-            user.EmailConfirmed = confirmed;
-            return Task.FromResult(0);
-        }
-
-        public async Task<ApplicationUser> FindByEmailAsync(string normalizedEmail, CancellationToken cancellationToken)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                await connection.OpenAsync(cancellationToken);
-                return await connection.QuerySingleOrDefaultAsync<ApplicationUser>($@"SELECT * FROM [ApplicationUser]
-                    WHERE [NormalizedEmail] = @{nameof(normalizedEmail)}", new { normalizedEmail });
-            }
-        }
-
-        public Task<string> GetNormalizedEmailAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            return Task.FromResult(user.NormalizedEmail);
-        }
-
-        public Task SetNormalizedEmailAsync(ApplicationUser user, string normalizedEmail, CancellationToken cancellationToken)
-        {
-            user.NormalizedEmail = normalizedEmail;
-            return Task.FromResult(0);
-        }
-
-        public Task SetPhoneNumberAsync(ApplicationUser user, string phoneNumber, CancellationToken cancellationToken)
-        {
-            user.PhoneNumber = phoneNumber;
-            return Task.FromResult(0);
-        }
-
-        public Task<string> GetPhoneNumberAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            return Task.FromResult(user.PhoneNumber);
-        }
-
-        public Task<bool> GetPhoneNumberConfirmedAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            return Task.FromResult(user.PhoneNumberConfirmed);
-        }
-
-        public Task SetPhoneNumberConfirmedAsync(ApplicationUser user, bool confirmed, CancellationToken cancellationToken)
-        {
-            user.PhoneNumberConfirmed = confirmed;
-            return Task.FromResult(0);
-        }
-
-        public Task SetTwoFactorEnabledAsync(ApplicationUser user, bool enabled, CancellationToken cancellationToken)
-        {
-            user.TwoFactorEnabled = enabled;
-            return Task.FromResult(0);
-        }
-
-        public Task<bool> GetTwoFactorEnabledAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            return Task.FromResult(user.TwoFactorEnabled);
-        }
-
-        public Task SetPasswordHashAsync(ApplicationUser user, string passwordHash, CancellationToken cancellationToken)
-        {
-            user.PasswordHash = passwordHash;
-            return Task.FromResult(0);
-        }
-
-        public Task<string> GetPasswordHashAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            return Task.FromResult(user.PasswordHash);
-        }
-
-        public Task<bool> HasPasswordAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            return Task.FromResult(user.PasswordHash != null);
-        }
-
-        public async Task AddToRoleAsync(ApplicationUser user, string roleName, CancellationToken cancellationToken)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                await connection.OpenAsync(cancellationToken);
-                var normalizedName = roleName.ToUpper();
-                var roleId = await connection.ExecuteScalarAsync<int?>($"SELECT [ApplicationUser_ID] FROM [ApplicationRole] WHERE [NormalizedName] = @{nameof(normalizedName)}", new { normalizedName });
-                if (!roleId.HasValue)
-                    roleId = await connection.ExecuteAsync($"INSERT INTO [ApplicationRole]([Name], [NormalizedName]) VALUES(@{nameof(roleName)}, @{nameof(normalizedName)})",
-                        new { roleName, normalizedName });
-
-                await connection.ExecuteAsync($"IF NOT EXISTS(SELECT 1 FROM [ApplicationUserRole] WHERE [User_ID] = @userId AND [] = @{nameof(roleId)}) " +
-                    $"INSERT INTO [ApplicationUserRole]([User_ID], [Role_ID]) VALUES(@userId, @{nameof(roleId)})",
-                    new { userId = user.ApplicationUser_ID, roleId });
-            }
-        }
-
-        public async Task RemoveFromRoleAsync(ApplicationUser user, string roleName, CancellationToken cancellationToken)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                await connection.OpenAsync(cancellationToken);
-                var roleId = await connection.ExecuteScalarAsync<int?>("SELECT [ApplicationUser_ID] FROM [ApplicationRole] WHERE [NormalizedName] = @normalizedName", new { normalizedName = roleName.ToUpper() });
-                if (!roleId.HasValue)
-                    await connection.ExecuteAsync($"DELETE FROM [ApplicationUserRole] WHERE [User_ID] = @userId AND [Role_ID] = @{nameof(roleId)}", new { userId = user.ApplicationUser_ID, roleId });
-            }
-        }
-
-        public async Task<IList<string>> GetRolesAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                await connection.OpenAsync(cancellationToken);
-                var queryResults = await connection.QueryAsync<string>("SELECT r.[Name] FROM [ApplicationRole] r INNER JOIN [ApplicationUserRole] ur ON ur.[Role_ID] = r.ApplicationRole_ID " +
-                    "WHERE ur.User_ID = @userId", new { userId = user.ApplicationUser_ID });
-
-                return queryResults.ToList();
-            }
-        }
-
-        public async Task<bool> IsInRoleAsync(ApplicationUser user, string roleName, CancellationToken cancellationToken)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                var roleId = await connection.ExecuteScalarAsync<int?>("SELECT [ApplicationUser_ID] FROM [ApplicationRole] WHERE [NormalizedName] = @normalizedName", new { normalizedName = roleName.ToUpper() });
-                if (roleId == default(int)) return false;
-                var matchingRoles = await connection.ExecuteScalarAsync<int>($"SELECT COUNT(*) FROM [ApplicationUserRole] WHERE [User_ID] = @userId AND [Role_ID] = @{nameof(roleId)}",
-                    new { userId = user.ApplicationUser_ID, roleId });
-
-                return matchingRoles > 0;
-            }
-        }
-
-        public async Task<IList<ApplicationUser>> GetUsersInRoleAsync(string roleName, CancellationToken cancellationToken)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                var queryResults = await connection.QueryAsync<ApplicationUser>("SELECT u.* FROM [ApplicationUser] u " +
-                    "INNER JOIN [ApplicationUserRole] ur ON ur.[UserId] = u.[ApplicationUser_ID] INNER JOIN [ApplicationRole] r ON r.[ApplicationUser_ID] = ur.[RoleId] WHERE r.[NormalizedName] = @normalizedName",
-                    new { normalizedName = roleName.ToUpper() });
-
-                return queryResults.ToList();
-            }
-        }
-
-        public void Dispose()
-        {
-            // Nothing to dispose.
-        }
-
-        public Task SetAuthenticatorKeyAsync(ApplicationUser user, string key, CancellationToken cancellationToken)
-        {
-            user.AuthenticatorKey = key;
-            return Task.CompletedTask;
-        }
-
-        public Task<string> GetAuthenticatorKeyAsync(ApplicationUser user, CancellationToken cancellationToken)
-        {
-            return Task.FromResult(user.AuthenticatorKey);
-        }
-    }
-}
