--- Museum.App.Services/Implementation/Servises/CategoryService.cs
@@ -1,15 +1,0 @@
-﻿using Museum.App.Services.Attributes;
-using Museum.App.Services.Interfaces.Servises;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Museum.App.Services.Implementation.Services
-{
-    [Service]
-    public class CategoryService : ICategoryService
-    {
-    }
-}
