+++ Museum.App.Services/Implementation/Servises/UserService/EmailSenderService.cs
@@ -1,0 +1,14 @@
+﻿using Museum.App.Services.Attributes;
+using Museum.App.Services.Interfaces.Servises;
+
+namespace Museum.App.Services.Implementation.Servises.Other
+{
+    [Service]
+    public class EmailSenderService : IEmailSenderService
+    {
+        public Task SendEmailAsync(string email, string subject, string message)
+        {
+            return Task.CompletedTask;
+        }
+    }
+}
