+++ Museum.App.Services/Interfaces/Servises/IFilterService.cs
@@ -4,15 +4,17 @@ namespace Museum.App.Services.Interfaces.Servises
 {
     public interface IFilterService
     {
-        public IEnumerable<FilterSectionViewModel> GetGalleryObjectsAsFilterPage();
         public IEnumerable<SideBarCollection> GetSideBarCollections();
-        public Task AddVote(VoteViewModel vote);
-        public Task RemoveVote(VoteViewModel vote);
-        public Task<bool> IsVoteExistForUser(int userId);
 
-        public Task<int> GetLikeCount(int objectId);
-        public Task<int> GetDislikeCount(int objectId);
-        public Task<int> GetCommentsCount(int objectId);
+        public Task<IEnumerable<FilterSectionViewModel>> GetGalleryObjectsAsFilterPageAsync();
+
+        public Task AddVoteAsync(VoteViewModel vote);
+        public Task RemoveVoteAsync(VoteViewModel vote);
+        public Task<bool> IsVoteExistForUserAsync(int userId);
+
+        public Task<int> GetLikeCountAsync(int objectId);
+        public Task<int> GetDislikeCountAsync(int objectId);
+        public Task<int> GetCommentsCountAsync(int objectId);
 
     }
 }
