+++ Museum.App/Utilites/EnumerableHelper.cs
@@ -7,5 +7,14 @@
         {
             return source.Skip((page - 1) * pageSize).Take(pageSize);
         }
+
+        public static async Task<IEnumerable<TSource>> PaginationAsync<TSource>(
+        this IEnumerable<TSource> source, int page, int pageSize)
+        {
+            return await Task.Run(() =>
+            {
+                return source.Skip((page - 1) * pageSize).Take(pageSize);
+            });
+        }
     }
 }
