+++ Museum.App/Views/GalleryObject/Index.cshtml
@@ -92,3 +92,106 @@
 <!-- End gallery Area -->
 
 
+<!-- Start blog Area -->
+<section class="blog-area section-gap" id="blog">
+    <div class="container">
+        <div class="row d-flex justify-content-center">
+            <div class="menu-content pb-70 col-lg-8">
+                <div class="title text-center">
+                    <h1 class="mb-10">Latest From Our Blog</h1>
+                    <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut
+                        labore et dolore magna aliqua.</p>
+                </div>
+            </div>
+        </div>
+        <div class="row">
+            <div class="col-lg-3 col-md-6 single-blog">
+                <div class="thumb">
+                    <img class="img-fluid" src="img/b1.jpg" alt="">
+                </div>
+                <p class="date">10 Jan 2018</p>
+                <a href="#"><h4>Addiction When Gambling
+                    Becomes A Problem</h4></a>
+                <p>
+                    inappropriate behavior ipsum dolor sit amet, consectetur.
+                </p>
+                <div class="meta-bottom d-flex justify-content-between">
+                    <p><span class="lnr lnr-heart"></span> 15 Likes</p>
+                    <p><span class="lnr lnr-bubble"></span> 02 Comments</p>
+                </div>
+            </div>
+            <div class="col-lg-3 col-md-6 single-blog">
+                <div class="thumb">
+                    <img class="img-fluid" src="img/b2.jpg" alt="">
+                </div>
+                <p class="date">10 Jan 2018</p>
+                <a href="#"><h4>Addiction When Gambling
+                    Becomes A Problem</h4></a>
+                <p>
+                    inappropriate behavior ipsum dolor sit amet, consectetur.
+                </p>
+                <div class="meta-bottom d-flex justify-content-between">
+                    <p><span class="lnr lnr-heart"></span> 15 Likes</p>
+                    <p><span class="lnr lnr-bubble"></span> 02 Comments</p>
+                </div>
+            </div>
+            <div class="col-lg-3 col-md-6 single-blog">
+                <div class="thumb">
+                    <img class="img-fluid" src="img/b3.jpg" alt="">
+                </div>
+                <p class="date">10 Jan 2018</p>
+                <a href="#"><h4>Addiction When Gambling
+                    Becomes A Problem</h4></a>
+                <p>
+                    inappropriate behavior ipsum dolor sit amet, consectetur.
+                </p>
+                <div class="meta-bottom d-flex justify-content-between">
+                    <p><span class="lnr lnr-heart"></span> 15 Likes</p>
+                    <p><span class="lnr lnr-bubble"></span> 02 Comments</p>
+                </div>
+            </div>
+            <div class="col-lg-3 col-md-6 single-blog">
+                <div class="thumb">
+                    <img class="img-fluid" src="img/b4.jpg" alt="">
+                </div>
+                <p class="date">10 Jan 2018</p>
+                <a href="#"><h4>Addiction When Gambling
+                    Becomes A Problem</h4></a>
+                <p>
+                    inappropriate behavior ipsum dolor sit amet, consectetur.
+                </p>
+                <div class="meta-bottom d-flex justify-content-between">
+                    <p><span class="lnr lnr-heart"></span> 15 Likes</p>
+                    <p><span class="lnr lnr-bubble"></span> 02 Comments</p>
+                </div>
+            </div>
+        </div>
+    </div>
+</section>
+<!-- End blog Area -->
+
+<section class="gradient-custom">
+    <div class="container my-5 py-5">
+        <div class="row d-flex justify-content-center">
+            <div class="col-md-12">
+                <div class="card">
+                    <div class="card-body p-4">
+                        <h4 class="text-center pb-2">Nested comments section</h4>
+
+                        <div class="row">
+                            <div class="col">
+                                <div class="main-comment-area">
+
+                                    
+
+                                </div>
+                            </div>
+                        </div>
+
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</section>
+
