+++ Museum.App/Views/GalleryObject/_singleComment.cshtml
@@ -1,0 +1,44 @@
+﻿
+<div class="single-comment">
+    <div class="d-flex flex-start mt-4">
+        <!-- Icon -->
+        <img class="rounded-circle shadow-1-strong me-3"
+             src="https://mdbcdn.b-cdn.net/img/Photos/Avatars/img%20(10).webp"
+             alt="avatar"
+             width="50"
+             height="50"/>
+
+        <div class="flex-grow-1 flex-shrink-1">
+            <div>
+                <div class="d-flex justify-content-between align-items-center">
+                    <p class="mb-1">
+                        <!-- NAME -->
+                        Maria Smantha
+                        <span class="small">
+                            <!-- Date of comment-->
+                            - 2 hours ago
+                        </span>
+                    </p>
+                    <a href="#!" class="btn">
+                        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor"
+                             class="bi bi-reply-all-fil l" viewBox="0 0 16 16">
+                            <path d="M8.021 11.9 3.453 8.62a.719.719 0 0 1 0-1.238L8.021 4.1a.716.716 0 0 1 1.079.619V6c1.5 0 6 0 7 8-2.5-4.5-7-4-7-4v1.281c0 .56-.606.898-1.079.62z"/>
+                            <path d="M5.232 4.293a.5.5 0 0 1-.106.7L1.114 7.945a.5.5 0 0 1-.042.028.147.147 0 0 0 0 .252.503.503 0 0 1 .042.028l4.012 2.954a.5.5 0 1 1-.593.805L.539 9.073a1.147 1.147 0 0 1 0-1.946l3.994-2.94a.5.5 0 0 1 .699.106z"/>
+                        </svg>
+                        <span class="small"> reply</span>
+                    </a>
+                </div>
+                <p class="small mb-0">
+                    <!-- Comment text -->
+                    text
+                </p>
+            </div>
+
+            <!-- Replies Area -->
+            <div class="comment-replies-area">
+
+            </div>
+
+        </div>
+    </div>
+</div>
