+++ Museum.Models/Enums/GroupKeys.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Museum.Models.Enums
+{
+    public enum GroupKeys
+    {
+        None = 0,
+        Name,
+        Popular
+    }
+}
