+++ Museum.Models/Enums/ObjectDescType.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Museum.Models.Enums
+{
+    public enum ObjectDescType
+    {
+        String = 0,
+        Table
+    }
+}
