+++ Museum.Models/GalleryObjectModels/CommentSectionModel.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Museum.Models.GalleryObjectModels
+{
+    public class CommentSectionModel
+    {
+
+        public string Comment_ID { get; set; }
+        public string CommentText { get; set; }
+        public string CommentLevel { get; set; }
+        public string CommentDate { get; set; }
+    }
+}
