--- Museum.Models/GalleryObjectModels/GalleryUlViewModel.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Museum.Models.GalleryObjectModels
-{
-    public class GalleryUlViewModel
-    {
-        public string? Title { get; set; }
-        public DateTime Create_Date { get; set; }
-        public string? Location { get; set; }
-        public string? Depositary { get; set; }
-        public string? Description { get; set; }
-    }
-}
